+++ src-tauri/src/ipc/get.rs
@@ -1,6 +1,7 @@
+use crate::database::DbPoolWrapper;
 use crate::hyprpaper;
 use crate::ipc::Response;
-use crate::{db_models, schema, DbPoolWrapper};
+use crate::{db_models, schema};
 use diesel::prelude::*;
 use tauri::State;
